manually fix some of the require field fix as described here #104
@teixeirak, I decided it was simpler to make the fixes manually in the form.
The only thing is that the ones for " if alive, and if % crown living - % crown intact<5, apply the % crown intact to both fields (n=3) " will show up as "missing_percent_crownt_intact" but I'll ignore.

And for the missing stems, unless I add lines to the form, I can't fix this since there are no lines for those. I'll see what happens if I leave it as is. I may end up doing a "auto fix" or just accept NA in the status.